rename gui module
