+++ mtgogui/src/layout.rs
@@ -1,0 +1,1 @@
+mod setup;
